Added edits from @gaurav-wustl that add the umbrella sampling functionality. This version builds off of edits from @minafarag1

+++ src/parsekey.h
@@ -12,8 +12,8 @@ int Parse_EnergyFile(char* strEnFile);
 
 void Parse_StructureFile(char* filename);
 
-void Parse_StructureFile_CalcBeadsAndChains(char* filename, int* n_bead_num, int* n_chain_types,
-                                            int* n_chain_num);
+void Parse_StructureFile_CalcBeadsAndChains(char* filename, size_t* n_bead_num, size_t* n_chain_types,
+                                            size_t* n_chain_num);
 
 void CreateBeadsAndChains(size_t n_bead_num, size_t n_chain_num);
 #endif // _PARSEKEY_H_
